pet hadLost & haFound fields added for statistic

+++ src/lost-info/lost-info.controller.ts
@@ -12,16 +12,22 @@ import {
   ValidationPipe,
   UseGuards,
   UnauthorizedException,
+  Query,
 } from '@nestjs/common';
 import { AuthGuard } from 'src/auth/guard/auth.guard';
 
 import { LostInfoService } from './lost-info.service';
+import { PetService } from 'src/pet/service/pet.service';
 
 import { CreateLostInfoDto } from './dto/create-lost-info.dto';
+import { QueryLostInfoDto } from './dto/query-lost-info.dto';
 
 @Controller('lost-info')
 export class LostInfoController {
-  constructor(private readonly lostInfoService: LostInfoService) {}
+  constructor(
+    private readonly lostInfoService: LostInfoService,
+    private readonly petService: PetService,
+  ) {}
 
   @UseGuards(AuthGuard)
   @Post(':petId')
@@ -30,9 +36,20 @@ export class LostInfoController {
     @Param('petId', ParseIntPipe) petId: number,
     @Body(new ValidationPipe({ transform: true, whitelist: true }))
     data: CreateLostInfoDto,
+    @Query(new ValidationPipe({ transform: true, whitelist: true }))
+    query: QueryLostInfoDto,
   ) {
     const tokenDecoded = req['user'] as TokenDecoded | undefined;
     if (!tokenDecoded) throw new UnauthorizedException();
+
+    if (query.addToStatistic) {
+      const [, err] = await this.petService.update({
+        where: { id: petId },
+        data: { hadLost: true },
+      });
+
+      if (err) throw err;
+    }
 
     const [lostInfo, err] = await this.lostInfoService.upsert({
       where: { petId, pet: { userId: tokenDecoded.sub } },
@@ -48,9 +65,20 @@ export class LostInfoController {
   async delete(
     @Req() req: Request,
     @Param('petId', ParseIntPipe) petId: number,
+    @Query(new ValidationPipe({ transform: true, whitelist: true }))
+    query: QueryLostInfoDto,
   ) {
     const tokenDecoded = req['user'] as TokenDecoded | undefined;
     if (!tokenDecoded) throw new UnauthorizedException();
+
+    if (query.addToStatistic) {
+      const [, err] = await this.petService.update({
+        where: { id: petId },
+        data: { hadFound: true, hadLost: true },
+      });
+
+      if (err) throw err;
+    }
 
     const [, err] = await this.lostInfoService.delete({
       where: { petId, pet: { userId: tokenDecoded.sub } },
